Revert the Pay page redesign (rolodex + credits block)
Reverts 758842a and 38b5ce0. The new layout didn't work visually; taking the page
back to its previous state rather than leaving a half-liked design in production
while we redesign it properly.

Deliberately KEPT:
- Phase 1 (KYC ungating) — behaviour, not design. openPortals stays ungated so bill
  management is still open to unverified members.
- Phase 4 — the card "coming soon" state and the Accounts deposit-account card,
  which are on a different surface and were separately requested.
